Add 'index' parameter to search function
The 'search' function in the client module of the 'vreedb' package has been updated. An 'index' parameter has been added to accommodate queries by index in addition to the existing parameters. The accompanying request payload formation in the asynchronous post method invocation is also updated to include the new 'index' data.