Changed async methods to synchronous in client.py
The changes in this commit move away from asynchronous I/O in the handling of HTTP client requests in the vreedb client. The AsyncClient from httpx is replaced by the Client class for synchronous requests, and all async keywords for function definitions have been removed. Response handling is also adjusted for synchronous operation.