Add max_distance_percent in search function
The search function in the client.py file has been updated to include a 'max_distance_percent' parameter. This parameter allows users to specify the maximum desired distance percentage during their search. This extends the functionality of the search and brings more control to the user in filtering their search results.